backend/pyproject.toml
[project]
name = "shelfready-backend"
version = "0.1.0"
description = "ShelfReady - AI Product Listing & Creative Engine for E-commerce"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "httpx>=0.27.0",
    "anthropic>=0.39.0",
    "supabase>=2.9.0",
    "stripe>=11.0.0",
    "resend>=2.0.0",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.7.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
